Fix memory leaks.
We use official SRS tetromino, and will not use other poliomino collections.
So it's not necessary to keep the possibilty to use different set of
poliominos.

Remove TetrominoCollection struct, as all rotations are now in a global
constant array.

Add init.c and init.h to intitialize tetramino colors which are not constant,
but as struct.

Use use/destroy for all struct, except for PositionInGrid, which is so simple
(two size_t) that we can pass it by value.

Now Piece store only the tetromino type (enum) to index the global rotation
and color arrays.